requirements.txt
backcall==0.2.0
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
cycler==0.10.0
decorator==5.0.9
dill==0.3.4
flatbuffers==2.0
idna==2.10
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
numpy==1.20.3
onnx==1.9.0
onnxruntime==1.8.1
opencv-python==4.5.2.54
pandas==1.2.5
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
protobuf==3.17.3
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
pywin32==301
PyYAML==5.4.1
pyzmq==22.1.0
requests==2.25.1
scikit-learn==0.24.2
scipy==1.7.0
seaborn==0.11.1
six==1.16.0
sklearn==0.0
threadpoolctl==2.2.0
timm==0.4.9
torch==1.8.1+cu102
torchaudio==0.8.1
torchinfo==0.1.4
torchvision==0.9.1+cu102
tornado==6.1
tqdm==4.61.2
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
